simd as separate crate
