+++ lib/simd/avx/src/avx_manhattan.rs
@@ -1,20 +1,17 @@
 use std::arch::x86_64::*;
 
-use common::types::ScoreType;
-
-use crate::data_types::vectors::VectorElementTypeByte;
-use crate::spaces::simple_avx::hsum256_ps_avx;
-
 #[target_feature(enable = "avx")]
 #[target_feature(enable = "fma")]
 #[allow(unused)]
-pub unsafe fn avx_euclid_similarity_bytes(
-    v1: &[VectorElementTypeByte],
-    v2: &[VectorElementTypeByte],
-) -> ScoreType {
+pub unsafe fn avx_manhattan_similarity_bytes(
+    v1: &[u8],
+    v2: &[u8],
+) -> f32 {
+    use crate::hsum256_ps_avx;
+
     debug_assert!(v1.len() == v2.len());
-    let mut ptr1: *const VectorElementTypeByte = v1.as_ptr();
-    let mut ptr2: *const VectorElementTypeByte = v2.as_ptr();
+    let mut ptr1: *const u8 = v1.as_ptr();
+    let mut ptr2: *const u8 = v2.as_ptr();
 
     // sum accumulator for 8x32 bit integers
     let mut acc = _mm256_setzero_si256();
@@ -30,48 +27,22 @@ pub unsafe fn avx_euclid_similarity_bytes(
         ptr1 = ptr1.add(32);
         ptr2 = ptr2.add(32);
 
-        // Compute the difference in both directions and take the maximum for abs
-        let diff1 = _mm256_subs_epu8(p1, p2);
-        let diff2 = _mm256_subs_epu8(p2, p1);
-
-        let abs_diff = _mm256_max_epu8(diff1, diff2);
-        let masked_abs_diff = _mm256_and_si256(abs_diff, mask_epu16_epu8);
-
-        // take from lane p1 and p2 parts (using bitwise AND):
-        // p1 = [byte0, byte1, byte2, byte3, ..] -> [0, byte1, 0, byte3, ..]
-        // p2 = [byte0, byte1, byte2, byte3, ..] -> [0, byte1, 0, byte3, ..]
-        // and calculate 16bit multiplication with taking lower 16 bits
-        // wa can use signed multiplication because sign bit is always 0
-        let mul16 = _mm256_mullo_epi16(masked_abs_diff, masked_abs_diff);
-
-        acc = _mm256_add_epi32(acc, _mm256_and_si256(mul16, mask_epu32_epu16));
-        let mul16 = _mm256_bsrli_epi128(mul16, 2);
-        acc = _mm256_add_epi32(acc, _mm256_and_si256(mul16, mask_epu32_epu16));
-
-        // shift right by 1 byte for p1 and p2 and repeat previous steps
-        let abs_diff = _mm256_bsrli_epi128(abs_diff, 1);
-        let masked_abs_diff = _mm256_and_si256(abs_diff, mask_epu16_epu8);
-
-        let mul16 = _mm256_mullo_epi16(masked_abs_diff, masked_abs_diff);
-
-        acc = _mm256_add_epi32(acc, _mm256_and_si256(mul16, mask_epu32_epu16));
-        let mul16 = _mm256_bsrli_epi128(mul16, 2);
-        acc = _mm256_add_epi32(acc, _mm256_and_si256(mul16, mask_epu32_epu16));
+        let sad = _mm256_sad_epu8(p1, p2);
+        acc = _mm256_add_epi32(acc, sad);
     }
 
     let mul_ps = _mm256_cvtepi32_ps(acc);
     let mut score = hsum256_ps_avx(mul_ps);
 
-    let remainder = len % 32;
+    let mut remainder = len % 32;
     if remainder != 0 {
         let mut remainder_score = 0;
-        for _ in 0..remainder {
+        for _ in 0..len % 32 {
             let v1 = *ptr1 as i32;
             let v2 = *ptr2 as i32;
             ptr1 = ptr1.add(1);
             ptr2 = ptr2.add(1);
-            let diff = v1 - v2;
-            remainder_score += diff * diff;
+            remainder_score += (v1 - v2).abs();
         }
         score += remainder_score as f32;
     }
@@ -82,20 +53,28 @@ pub unsafe fn avx_euclid_similarity_bytes(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::spaces::metric::Metric;
-    use crate::spaces::simple::*;
+
+    fn manhattan_similarity_bytes(
+        v1: &[u8],
+        v2: &[u8],
+    ) -> f32 {
+        -v1.iter()
+            .zip(v2)
+            .map(|(a, b)| (*a as i32 - *b as i32).abs())
+            .sum::<i32>() as f32
+    }
 
     #[test]
     fn test_spaces_avx() {
         if is_x86_feature_detected!("avx") && is_x86_feature_detected!("fma") {
-            let v1: Vec<VectorElementTypeByte> = vec![
+            let v1: Vec<u8> = vec![
                 255, 255, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 255, 255,
                 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 255, 255, 0, 1, 2, 3,
                 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 255, 255, 0, 1, 2, 3, 4, 5, 6, 7,
                 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 255, 255, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10,
                 11, 12, 13, 14, 15, 16, 17,
             ];
-            let v2: Vec<VectorElementTypeByte> = vec![
+            let v2: Vec<u8> = vec![
                 255, 255, 0, 254, 253, 252, 251, 250, 249, 248, 247, 246, 245, 244, 243, 242, 241,
                 240, 239, 238, 255, 255, 255, 254, 253, 252, 251, 250, 249, 248, 247, 246, 245,
                 244, 243, 242, 241, 240, 239, 238, 255, 255, 255, 254, 253, 252, 251, 250, 249,
@@ -105,8 +84,8 @@ mod tests {
                 240, 239, 238,
             ];
 
-            let dot_simd = unsafe { avx_euclid_similarity_bytes(&v1, &v2) };
-            let dot = <EuclidMetric as Metric<VectorElementTypeByte>>::similarity(&v1, &v2);
+            let dot_simd = unsafe { avx_manhattan_similarity_bytes(&v1, &v2) };
+            let dot = manhattan_similarity_bytes(&v1, &v2);
             assert_eq!(dot_simd, dot);
         } else {
             println!("avx test skipped");
