+++ lib/simd/neon/src/neon_cosine.rs
