+++ lib/simd/neon/src/neon_euclid.rs
