+++ lib/simd/neon/src/neon_simple_manhattan.rs
