+++ lib/simd/sse/src/sse_manhattan.rs
@@ -1,19 +1,16 @@
 use std::arch::x86_64::*;
 
-use common::types::ScoreType;
-
-use crate::data_types::vectors::VectorElementTypeByte;
-use crate::spaces::simple_sse::hsum128_ps_sse;
-
 #[target_feature(enable = "sse")]
 #[allow(unused)]
 pub unsafe fn sse_manhattan_similarity_bytes(
-    v1: &[VectorElementTypeByte],
-    v2: &[VectorElementTypeByte],
-) -> ScoreType {
+    v1: &[u8],
+    v2: &[u8],
+) -> f32 {
+    use crate::hsum128_ps_sse;
+
     debug_assert!(v1.len() == v2.len());
-    let mut ptr1: *const VectorElementTypeByte = v1.as_ptr();
-    let mut ptr2: *const VectorElementTypeByte = v2.as_ptr();
+    let mut ptr1: *const u8 = v1.as_ptr();
+    let mut ptr2: *const u8 = v2.as_ptr();
 
     // sum accumulator for 4x32 bit integers
     let mut acc = _mm_setzero_si128();
@@ -55,20 +52,28 @@ pub unsafe fn sse_manhattan_similarity_bytes(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::spaces::metric::Metric;
-    use crate::spaces::simple::*;
+
+    fn manhattan_similarity_bytes(
+        v1: &[u8],
+        v2: &[u8],
+    ) -> f32 {
+        -v1.iter()
+            .zip(v2)
+            .map(|(a, b)| (*a as i32 - *b as i32).abs())
+            .sum::<i32>() as f32
+    }
 
     #[test]
     fn test_spaces_sse() {
         if is_x86_feature_detected!("sse") {
-            let v1: Vec<VectorElementTypeByte> = vec![
+            let v1: Vec<u8> = vec![
                 255, 255, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 255, 255,
                 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 255, 255, 0, 1, 2, 3,
                 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 255, 255, 0, 1, 2, 3, 4, 5, 6, 7,
                 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 255, 255, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10,
                 11, 12, 13, 14, 15, 16, 17,
             ];
-            let v2: Vec<VectorElementTypeByte> = vec![
+            let v2: Vec<u8> = vec![
                 255, 255, 0, 254, 253, 252, 251, 250, 249, 248, 247, 246, 245, 244, 243, 242, 241,
                 240, 239, 238, 255, 255, 255, 254, 253, 252, 251, 250, 249, 248, 247, 246, 245,
                 244, 243, 242, 241, 240, 239, 238, 255, 255, 255, 254, 253, 252, 251, 250, 249,
@@ -79,7 +84,7 @@ mod tests {
             ];
 
             let dot_simd = unsafe { sse_manhattan_similarity_bytes(&v1, &v2) };
-            let dot = <ManhattanMetric as Metric<VectorElementTypeByte>>::similarity(&v1, &v2);
+            let dot = manhattan_similarity_bytes(&v1, &v2);
             assert_eq!(dot_simd, dot);
         } else {
             println!("avx test skipped");
